Add Clean Arhitecture
This commit changes the structure of the project to follow Clean Arhitecture:

- VacationModule.API takes care of the controllers

- VacationModule.Core takes care of the business part

- VacationModule.Infrastructure takes care of the database

- VacationModule.IntegrationTests takes care of the integration tests

- VacationModule.ServiceTests takes care of the unit tests

The benefit of implementing Clean Arhitecture is that, unlike with N-Tier Arhitecture, the bussines logic is independent. It does not depend on the Data Acces Layer (Infrastructure). The dependency is inversed, so we improve scalability, testing and also in the future we can change the default SqlServer connection. To my knowledge this arhitecture is recommended to large scale applications and I will implement this project using it for learning purposes.